code refactoring and tests files setup
