+++ requirements.txt
@@ -27,3 +27,6 @@ duckdb
 sqlalchemy # needed for Ibis + DuckDB
 pyarrow
 pyarrow-hotfix
+pytest
+playwright
+pytest-playwright
