+++ src/app.py
@@ -17,54 +17,7 @@ from src.ui import app_ui
 from src.plot import build_temp_chart, build_yearly_plot, build_diff_plot
 from src.data_count import data_count_prep
 from src.map import build_base_map, apply_country_highlight
-
-
-def _diverging_styles(cells: list[tuple[int, int, float]], alpha: float = 0.65) -> list[dict]:
-    """
-    Generate DataGrid styles for a diverging color scale (0=neutral, negative=blue, positive=red).
-    cells: list of (row_idx, col_idx, value)
-    alpha: background opacity (higher = deeper colors)
-    """
-    if not cells:
-        return []
-
-    vals = [v for _, _, v in cells]
-    low = min(min(vals), 0)
-    high = max(max(vals), 0)
-    eps = 1e-9
-    if high - low < eps:
-        high = low + eps
-
-    WHITE = (255, 255, 255)
-    BLUE = (41, 128, 185)
-    RED = (231, 76, 60)
-
-    def _interp(t: float, c0: tuple, c1: tuple) -> str:
-        r = int(c0[0] + t * (c1[0] - c0[0]))
-        g = int(c0[1] + t * (c1[1] - c0[1]))
-        b = int(c0[2] + t * (c1[2] - c0[2]))
-        return f"rgba({r},{g},{b},{alpha})"
-
-    styles = []
-    for row_idx, col_idx, val_num in cells:
-        if val_num < 0:
-            t = val_num / low if low < -eps else 0
-            t = max(0, min(1, t))
-            bg = _interp(t, WHITE, BLUE)
-        elif val_num > 0:
-            t = val_num / high if high > eps else 0
-            t = max(0, min(1, t))
-            bg = _interp(t, WHITE, RED)
-        else:
-            bg = "rgba(248,249,250,0.5)"
-
-        styles.append({
-            "rows": [row_idx],
-            "cols": [col_idx],
-            "style": {"color": "#333", "backgroundColor": bg}
-        })
-    return styles
-
+from src.table_styles import diverging_styles, table_styles_wide
 
 def server(input: Inputs, output: Outputs, session: Session):
 
@@ -346,7 +299,7 @@ def server(input: Inputs, output: Outputs, session: Session):
                         cells.append((i, change_col, float(val)))
                     except (ValueError, TypeError):
                         pass
-        styles = _diverging_styles(cells)
+        styles = diverging_styles(cells)
         return render.DataGrid(df_table, selection_mode="none", styles=styles)
 
     # =============================
@@ -371,25 +324,6 @@ def server(input: Inputs, output: Outputs, session: Session):
     # =============================
     # Data Table (Monthly Comparison)
     # =============================
-    def _table_styles_wide(df: pd.DataFrame):
-        """Diverging gradient for Change row (row 2): 0=neutral, negative=blue, positive=red."""
-        if df.empty or df.shape[0] < 3:
-            return []
-        row_idx = 2
-        cells = []
-        for col_idx in range(1, df.shape[1]):
-            try:
-                val = df.iloc[row_idx, col_idx]
-            except (IndexError, KeyError):
-                continue
-            if pd.isna(val):
-                continue
-            try:
-                cells.append((row_idx, col_idx, float(val)))
-            except (ValueError, TypeError):
-                continue
-        return _diverging_styles(cells)
-
     @render.data_frame
     def data_table():
         """Table of monthly comparison in wide format (3 rows x 12 columns)."""
@@ -407,7 +341,7 @@ def server(input: Inputs, output: Outputs, session: Session):
                 selection_mode="none",
                 height="auto"
             )
-        return render.DataGrid(data, selection_mode="none", styles=_table_styles_wide, height="auto")
+        return render.DataGrid(data, selection_mode="none", styles=table_styles_wide, height="auto")
 
     @render.download(filename=lambda: _csv_download_filename())
     def download_table_csv():
